Make chunk sizes configurable (binary compat break)
Introduce a new command line option (--chunk-size-avg) that allows setting
the average chunk size. The implementation will then try and find the
closest prime number to the given value, and derive the absolute minumums and
maximums from that.

Note that this needs to be set to the same value for all invocations for
now, otherwise the chunker will not be able to detect any existing content.