casync-tool: add existing file as seed during extract
When extracting, and the file passed as output file does already exist, use
it as extra seed automatically. This way, a file can be extracted, and only
parts that are missing are taken from the store. If the store is remote,
this will save network bandwidth.

Maybe that behavior can be made optional in the future.