caencoder: optionally respect the -d chattr(1) flag, and exclude sysfs/procfs and co from serialization
By default we'll now respect the FS_NODUMP_FL for selecting what to
include and what not in the serialization. This may be turned off with
--respect-nodump=no however.

Also, let's ignore all API VFS, such as procfs, sysfs and friends.

Note that for the latter we still store an empty dir, for the former
nothing at all. That's because it's a good idea to retain mount points
as mount points