casync: when calculating digest of device node, consider this "encoding"
For us, the encoding of a regular file or device node is equal to the
contents of the file or device. This means, when calculating the digest
of them it should not matter if we use the "encoder" logic for it or the
"decoder", the result should be the same.

Except that when encoding a device node, we actually try to read the
real file system size of any stored superblock, and that's a logic we
want to take benefit of it in that case. We don't do this when decoding,
though, hence prefer encoding over decoding in this case.