casync: rework serialization format
this splits out the filename from the ENTRY rercord, into a new FILENAME
record. This allows us to store the file attributes of the drectory
root, for which we only have file metadata but no filename.

As side effect this removes HELLO records, as each stream now begins
with an ENTRY record.

The feature flag set is therefore moved into to ENTRY.

BREAKS FILE FORMAT COMPATIBILITY!

(Note that this also drops the "second part" of the UUID in the format,
as it appears too much and unnecessary to add this to each ENTRY record)