casync: rework state engine, to add an extra state when done with each file
This adds a set of events CA_ENCODER_DONE_FILE, CA_DECODER_DONE_FILE and
CA_SYNC_DONE_FILE which are returned whenever the state machines
finished processing a file. This state may be used as local EOF marker.

"casync mtree" now makes use of this to show the digest for eahc file
after is is complete.

The "casync" tool will now process events with the same generic handler
function if possible, shortening the code a bit.

Fixes: #13