casync: store feature flag in index and use it for seeds
This makes use of the existing "feature flags" field in index files and
uses it to store the same feature flags as used for archive generation
in them. This is then used during seeding to ensure we get the right
seed data set.

This simply extends the existing logic for propagating chunk sizes to
the feature flags.