casync: rework cross-device logic to be marked in feature flags field
This reworks the --xdev= option to become --exclude-submounts=yes|no,
and the --respect-nodump= option to become --exclude-nodump=yes|no, thus
streamlining the user interface. This is the also reflected in the
archive feature flags field, so that both settings are stored there.
This is useful so that an index file contains a complete description of
what needs to be included in the archive and what not, for seeding
purposes.