casync: issue READY=1 through sd_notify() when casync finished initialization
This way "casync mount" and "casync mkdev" will issue a ready
notification as soon as they established the device or mount point, so
that clients can safely synchronize against that. A test tool
"notify-wait" is also added (and used by test-nbd.sh and test-fuse.sh)
for waiting for READY=1 on arbitrary commands forked into the
background.