cachunker: replace adler32 with buzhash
buzhash is a bit faster and returns more evenly distributed results,
which makes our chunking more deterministic.

This also changes the calculation of the cut decision, so that the
average chunk size actually matches what is selected.

The default chunk size is now increased to 64K.

Finally, "--chunk-size-avg=" is replaced by "--chunk-size=" which either
takes a single average chunk size or a triplet of mininimal, average and
maximum chunk size, separated by colons.

This change does not break the binary format, and previously generated
index files stay valid, but given that we chunk
differently seeding using a new casync will not be effective when
processing an old index file.