casync: rework index and archive file format to have recognizable endings
This alters the index and archive formats to have clearly recognizable
file endings. Specifically the following changes are made:

1. The index file's overall "digest" field a the end of the file is
   removed without replacement. It wasn't used or verified so far, and
   it makes it impossible to generate index files without reading the
   entire payload again, which is highly problematic when adapting
   casync for regular backups without requiring full reading of the
   entire dataset each time.

2. Both the GOODBYE object in archive files and the TABLE object in
   index files gained a "marker" (a constant 64bit field that may be
   used to recognize the end of archive and index files.

3. The GOODBYE object now has a field specifiying the offset from its
   beginning to the beginning of the matching ENTRY object. Adding this
   allows parsers to determine whether there's preceeding garbage in
   archive files, as for the top-level directory this offset must point
   to file offset 0.

4. Similar, TABLE objects now carry a pointer to the matching INDEX
   header, for similar reasons.

With all of this in place casync might eventually learn to deal with
archive files that have been exended to multiples of some sector size,
as it's not possible to search backwards from the extended end until the
marker with the right ENTRY offset is found.

THIS COMMIT BREAKS THE ARCHIVE AND INDEX FORMAT IN AN INCOMPATIBLE WAY!