util: simplify meson logic by dropping "struct file_clone_range" check
Let's not bother with checking in meson whether struct file_clone_range
is defined in linux/fs.h, let's instead just reuse the FICLONERANGE
macro definition as check, which should be equivalent, but allows us to
shorten our meson.build. file.

As a handy side-effect this closes #31 for us, even though this doesn't
actually fix the real problem, which appears to be a bug in Meson.

Fixes: #31