canbd: create nbd block device symlinks via udev, if available
This cleans up symlink management for NBD block devices. Specifically,
if the user invokes:

        casync mkdev somefile.raw /dev/foobar

Then the /dev/foobar symlink is created through udev, so that it is
properly registered in the udev database, may be queried from there and
is removed along with the device node.

This is implemented by storing the specified name in
/run/casync/$DEVICE and a small udev rule that reads it from there.

A file lock is held on the device to indicate whether the data is still
relevant, so that we don't create symlinks for long recycled NBD
devices.