casync: replace gcrypt dep by openssl dep
It appears that openssl is more commonly used these days, hence let's
replace the libgcrypt dependency by an OpenSSL one, to follow where
things gravitate towards.

This is particular useful as OpenSSL permits us to do SHA512/256 in
addition to SHA256, which is most likely a better choice for our
purpose, simply because it is faster on the usual machines. (Well,
neither OpenSSL nor libgcrypt actually support SHA512/256 natively,
however as OpenSSL opencodes SHA512 we can patch in the different start
values, and make things work for us, which libgcrypt doesn#t permit us
to do.)

While we are at it, rework the digest code and introduce a generic
CaDigest object that hides that OpenSSL is now used underneath. This way
our choice of library won't leak into the API, which is a major,
relevant step towards making the casync API public.