casync: use sha512/256 as hash function by default
sha256 remains supported, via --digest=sha256, as it has benefits when
focussing on 32bit systems, but the new default is sha512/256, as it is
substantially faster on 64bit systems.

The used hash algorithm is encoded in the feature flags field in the
index and archive files, so that directing casync onto index or archive
files is enough for it to determine how to encode/look for chunks.