casync: change suffix of compressed chunks to .cacnk
So far we used the suffix ".xz", but this is misleading, given that we
can now use a variety of compressors, not just xz. Hence change the
suffix to .cacnk. To maintain a limited level of compatibility, allow
overriding the suffix via the $CASYNC_COMRPESSED_CHUNK_SUFFIX
environment variable. Set it to ".xz" and compatibility with the old
naming is retained.