test-nbd: avoid false-positive due to nbd readiness
In automated tests depending on the environment several cases hit an error
which appears like:
  ValueError: Invalid file object: <_io.BufferedReader name=5>
  FAILED: meson-test

In the background of this timeout there is the dd of test-nbd
failing due to test-node not being ready yet. That makes the test script to
fail and end up a zombified process.

After some time meson cleans up as an error.
To avoid that issue retry the dd with an incrementing wait.
That means in most cases no slowdown at all, while in affected cases a
correct test execution.

fixes 113

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>