casync: add caching logic for generating indexes/chunks
This adds a new subsystem for speeding up repeated generation of index
files from a file system tree, for example to implement relatively
high-frequency backups.

At its essence this adds a "cache" subsystem: for every chunk generated
off a file system tree a cache item is generated, keyed by the location
in the tree, and returning the chunk's hash value, plus a serialization
of the locations of all file system objects contributing to the chunk.
This way, when generating an image it is possible every time a chunk is
completed (as well as at the beginning of the stream) to check the
cache: if it contains an entry there's a high likelihood that the stored
chunk hash can be reused. However, before this can be safely reused, the
contributing files need to be looked at: if they have been modified
since the cache entry was generated the cache entry is out-of-date and
cannot be used. Moreover, if files where inserted or removed the entry
cannot be used either. Validation of found cache entries is implemented
by continuing iteration through the file system tree similar how it
would be done for non-cache serialization, however the payload of the
files is skipped (thus, we only look at the stat() data and related):
the locations we look at are then compared with the series retrieved
from the cache.

If the cached series of locations doesn't match the series of locations
retrieved from disk then the cache item is considered invalid. In this
case the encoding logic seeks back to the location where it found the
cache item, and starts serialization again, this time in normal,
non-cache mode. When a chunk is completed the cache is then checked
again, and if something is found the validation scheme is turned on
again for this chunk, as above.

Some timing results based on a sample tree (all with hot kernel caches,
with no changes to the tree in between)

1) First invocation of "casync make --cache=/var/tmp/somedir test.caidx"
   with an unpopulated cache in /var/tmp/somedir: 6.70s

2) Second invocation of the same command with a populated cache from the
   first run: 0.15s

In order to implement this logic, the "CaLocation" object (which encodes
the state of an encoder engine at any time) is greatly extended:

1) It gained an mtime, inode and generation field: these fields are
   populated from the stat() data (as well as FS_IOC_GETVERSION), and
   are useful for identifying a specific version of a file or directory.
   When a file is updated or replaced at least one if this three fields
   should change too.

2) It gained a "name_table" object reference. This object encodes the
   filename table data of all parent directories of serialized objects.
   This is useful so that we can later return to a location and restore
   all filename offsets from the earlier run, for all directories
   referenced in the location, so that we later on can generate the
   GOODBYE filename offset tables properly.

3) It gained an "archive_offset". This contains the current encoder byte
   counter, which is also useful to later return to an existing
   location: we can start serialization exactly from the point where we
   left off, with the relevant counter back initialized.

The CaLocation object also gained a new ca_location_id_make() call which
generates a hash of the specified location, which we use as lookup key
when maintaining the persistant cache on disk.