cadecoder: update validity check for parent node when finalizing
Previously we'd access the "n" object at these places, but we don't do
this anymore, as we have cached "dir_fd" — which is the relevant field
we need — hence let's check that one instead. Also, we filtered out the
case when we have neither n->fd and child->fd already further above,
hence let's be more strict here now, and use an assert.