cadecoder: verify timestamp of file after finalizing it
Unfortunately the kernel doesn't let us know whether the timestamp
granularity of the backing fs we operate on can provide what we need.
Which is annoying, as we want to guarantee reproducability. Hence, let's
check the mtime on the file after we finalized a file, and see if it
matches our expectations. And if not, fail the decoding.