src/cadecoder: fix getpwnam_r() and getgrnam_r() return code handling
If for example a catar file contains a file whose owner is not existing
on the target system, getpwnam_r() will return 0 and set pw to NULL.

In the former return code handling, only r == 0 with pw *not* being NULL
was checked and properly handled.
For all other cases we fell into the error return conditional which only
checked for getting ERANGE and otherwise makes negative error numbers
from positive ones or returns -ESRCH for 'error' numbers being
non-negative.

Thus, for the case described above, we fell into the return code check
with having a non-negative return code '0' and thus simply returned
-ESRCH ('No such process') from name_to_uid() which is definitely the
wrong conclusion and leads to confusing error messages such as:

| Failed to run synchronizer: No such process

Instead we need an extra case to handle a successful call to
getpwnam_r() indicating that the entry we were looking for was not
found.

This patch resolved this for both name_to_uid() and name_to_gid() by
checking for pw == NULL as indication of unsuccessful call to
getpwnam_r() or getgrnam_r() with an additional conditional for
distinguishing an orderly return indicating no entry was found from a
real error.
This is also close to what the man page suggest as example code.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>