casync: when seeking in the index due to remote prefetch, seek back one
When remote prefetch is enabled we will temporarily seek around in the
index in order to to figure out what to download next. When returning to
the original position afterwards, go one item back and read it again.
That's necessary, so that we know the offset of the previous item again,
so that calculation of the chunk lengths (which is just the difference
between the end offset of the current chunk and the end offset of the
previous chunk) works properly again.

Also, add some debug logging to make it easier to find issues like this
one.

Fixes: #112