cadecoder: accept if user name is encoded for UID == 0
Normallly we suppress user name records for UID 0, given that it is
"root" anyway. This logic doesn't work however if UID shifting is
applied. In that case the user name encoded still is the one of the host
system, while the numeric UID is shifted. This means, that it might very
well be that UID 0 maps to a user != "root", and we should permit this
in the archive file.

This hence losens the restrictions in the decoding code a bit for this,
but the encoding side remains unchanged since it already did the right
thing, and suppressed the name based on the unshifted UID already,
correctly.