cadecoder: fix decoding of GOODBYE records after non-directory entries
We use GOODBYE and FILENAME records as indicator that a previous entry's
serialization is complete. The previous entry may be anything not just a
directory, hence remove a misplaced safety check. (The check was added
under the assumption that the entry we are looking at was the directory
entry we are in, but actually it's just the current entry within it we are
processing).