casync: implicitly synthesize "root" user name when necessary
We suppress generation of user name records in archives for UID 0 if the
numeric UID is encoded as well, and the user name matches "root" for
that. Let's make sure that when reading archives we hide this, and
synthesize the "root" username when necessary. After all the user really
shouldn't know of this implementation detail, but always see user names
when he asked for user names.