cadecoder: don't shift UID/GID when user queries them of current file
Currently, while decoding we'd return the host version of UID/GID and
while encoding thea arhive version of them, when the user asks for it.
This makes little sense. Let's always return the archive version, so
that the listings always reflect our native data, and the UID/GID shift
is applied as part of file system access only.