tool: refuse "casync stat" on a regular file
We only support directories (for encoding) or .catar/.caidx files (for
decoding), but we can't really do anything useful for other regular
files, hence let's print a brief error message and refuse.

Previously we'd treat the input as ".catar" file in this case, but this
is quite confusing, as "casync stat /etc/fstab" results in a decoding
error. Later on we should make commands like that do something useful
(for example, encode /etc, then seek to /etc/fstab), hence for now let's
block this call with an error message, to keep our options open.