caencoder: before generating xattr/fcaps records check flags again
Usually, we encode everything we know about the files we are processing,
and the flags checking is done at the moment of reading this information
in from the files, but not later anymore. This doesn't work for
xattr/fcaps however, since the data source for both is the same, and
hence when we read one we also read the other. Hence, as a special case
for these bits, check the flags not only when reading things in but also
when generating records.