caseed: when we want to store overly long location strings, use files
Previously we exclusively used symlinks for caching information
generated by seeds: the mapping of chunk ids to locations encoding how
to regenerate the requested data again. However, this is problematic if
the data to encode is getting overly long, for example when the path
gets too long. Moreover as we encode more data in the location string
htis becomes even more problematic.

Hence, let's fix this by using symlinks whenever we can, but adding a
fall-back to regular files if we can't.