seed: make sure we can regenerate GOODBYE records from seed too
Previously, seeding hadone major limitation: any chunk that contained a
GOODBYE record could not be recreated on request. That was because
GOODBYE records contain offset tables to the files contained in the
directory they are used for. Since we didn't include these offsets in the
location the seed stored, we hence also couldn't regenerate the GOODBYE
records properly and hence not the chunks containing them either.

With this change the name table and offset data is now optionally
included in the string version of a location, and thus can be
saved/restored in the seed's lookup table.

Since some users of CaLocation want the name table and archive offset
included in the formatted string version and others do not want this,
this adds a new flags type CaLocationWith which contains flags for
various bits of information to encode in the string version of a
CaLocation object. This flags set is then also used when comparing
locations, as sometimes it is needed to compare with only some subset of
the fields.