add support for excluding certain files and directories with ".caexclude"
This adds a logic for excluding files and directories from archives by
including ".caexclude" files in the source file tree. The concept is
inspired by git's .gitignore files, and is mostly compatible.

This is implemented as a new CaMatch object that contains rules read
from these files organized in a tree (that's because .caexclude files —
like .gitignore files – can include full multi-level paths). When
testing a file name against such an object we always implicitly generate
a new CaMatch object including all matches that apply to the specific
subtree so that matches are properly propagated down thre tree where
that's appropriate. Match trees can also be merged, which is used when a
.caexclude contains rules that also apply to some child directory, but
the directory also includes a .caexclude file of its own.

Whether .caexclude files are respected or not may be configured with a
new --exclude-file=yes|no command line option, but by default they are
respected. Whether the files are respected or not is also included in
the feature flags bitmask included in index files and archive files.
That's a good idea so that seeding can work correctly and it is
sufficient to pass an index file to casync and all required exclusion
settings can be derived correctly from it.

Since the CaMatch logic allows merging of exclusion match trees we can
reuse the concept later to permit configuration of exclusion lists on
the command line, too. However, when we do that we should make sure to
include these in the index and archive files too, in order not to
compromise the automatic seed logic for them either.

Also, the CaMatch logic is reusable enough to be be useful later on to
implement concepts such as per-file encryption settings (usecase: for
some files it might be wise to disable chunking locally, and for others
it might make sense to disable encryption altogether or enable secure
deduplication in order to increase the chance of sharing data chunks
with others).