caformat: don't mask out feature flags based on compile time options in public headers
The caxyz.h headers should really stay free of #ifdef-fery based on
compile time options, as it's the stuff that is supposed to become the
publicly installed library API headers.

Hence, let's define the feature bits unconditionally in the header in
all its most featureful flory, but then let's mask out what we can't
support internally.

This is a rework of #127