Add a simple script to report holes in files
We make various efforts to punch holes, but it's not really clear how good
the effects are. This simple script for example reports:

$ echo -n 'a' >zeros; truncate -s $((4096*4)) zeros; test/report-holes.py zeros
found hole between 0x00001000 and 0x00004000 (12288 bytes)

My results, with kernel 4.14.14-300.fc27.x86_64, that on ext4 and btrfs, holes
are always aligned to 4096 bytes and their size is multiple of that alignment.