realloc-buffer: add safety size check
Let's put a limit on what we want to allocate at maximum. Let's set it
to 1G for now.

This is a precaution only. If we ever hit this limit we probably hit a
bug somewhere else.

(Also, let's add some _unlikely_ decorators around a couple of safety
bound checks that should never be trigger IRL)