meson: drop -ffast-math from gcc options
We don't do much floating point processing, and given that #133 has
shown that the inaccuracies introduced by -ffast-math have an effect on
us, let's sidestep the whole issue and turn off the whole thing for us.

After all reproducibility is supposed to be a big thing in casync, hence
we should really come to the same results on all archs, and all
compilers and all versions thereof.

See: #133