casync-http: retry when curl fails to connect
This patch implements a retry in casync-http, in the case where
curl_easy_perform returns CURLE_COULDNT_CONNECT.

Issue (how to reproduce)
------------------------

Guest is a QEmu VM, host is Debian testing.

    $ qemu-system-x86_64 -enable-kvm \
        -nic user,hostfwd=tcp::2222-:22 \
        ...

Python3 http server is running on the host.

    $ python3 -m http.server
    Serving HTTP on 0.0.0.0 port 8000 (http://0.0.0.0:8000/) ...

CASync is used on the guest.

    $ casync --version
    casync 2
    $ apt-cache show casync | grep ^Version
    Version: 2+20180321-2

With this setup, 'casync extract' fails consistently with:

    $ casync extract http://10.0.2.2:8000/rootfs.img.caibx /dev/sda6
    Failed to acquire http://10.0.2.2:8000/default.castr/890d/890d...0c74a.cacnk
    Failed to run synchronizer: Broken pipe

Debug
-----

Investigation shows that the sub-process casync-http fails at some point,
curl_easy_perform() returns CURLE_COULDNT_CONNECT. This error is not
handled and the process exits.

Some more research seems to indicate that this error is not considered
as a transient error by curl developers, hence curl won't retry [1]. How
to handle that is left to the caller.

So it's up to casync to decide if it should retry automatically, or if
it should defer this decision one layer up to the caller.

I must say that the error here is not reproducible if I use a proper
http server on the host (namely nginx) instead of python http server.
However I think that casync should be more robust in this regard.

I could just blindly restart casync when it fails, however `casync
extract` starts by chunking its seeds, which is an expensive operation.
So it's much more efficient if casync can handle this error internally.

----

[1]: https://curl.haxx.se/mail/tracker-2012-02/0014.html

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>