Add exit signal handler to casync-http
When hitting "Ctrl-C" during `casync extract`, the casync-http worker
receives a signal which is not handled, resulting in an unclean exit:
the `/var/tmp/<...>.carem` directory is not deleted.

This commit adds basic signal handling for exit signals, ensuring clean
exit and deletion of the temporary work directory in `/var/tmp`.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>