Make the compressor libraries configurable
This makes use of the meson 'feature' feature, added in meson 0.47, hence the
meson version requirement is bumped.

By default, we still fail at configuration time if not all supported
dependencies are present, but the user may set -Dauto_features=auto to do
autotools-style autoconfiguration, and -Dauto_features=disabled to disable all
optional features (and most likely enable select ones on top).

If compression with a disabled compressor is requested, the user gets a nice
error:
$ casync make --compression=zstd ...
Compiled without support for zstd compression

If de-compression that requires a disabled compressor is necessary, the user
gets just a cryptic error:
Failed to acquire chunk: Function not implemented
Failed to run synchronizer: Function not implemented

Tests are skipped for features which are not implemented.