cachunker: optionally search for "cutmarks" while chunking
Let's optimize chunking a bit: let's optionally accept a list of
"cutmarks", which are special bit sequences (up to 64bit) that indicate
particularly good chunking cut points. This can be used to optimize
chunking in data streams we know a few semantics about.

The intention is that the object markers of .caidx files are set to be
cutmarks, so that we rather cut between objects than at entirely
arbitrary positions.

This commit only adds logic to find these cutpoints, based on a list of
defined cutpoints, the code calling into the chunker does not make use
of it yet.