tests: do not work on the whole "source" directory, which can be very large
As soon as I build casync for a few different architectures (each with
a separate build directory), or if I run some failing tests and get a core
dump in the "source" directory, the tests start to take a few minutes each.
This is pointless. Let's use a much smaller set of files for testing, for
directories where we don't expect any "random" files to be crated. The
tests still take a few seconds each on my laptop, so quite a bit of
work is being done.