shell-completion: add helper __casync_get_arg_n
Introduce helper function __casync_get_arg_n to return the n-th argument
in $arg. The function takes into consideration the options whose the
following argument should not be counted as argument, and the options
that should be counted as argument.

It replaces the helper function in_array to get the command from the
already entered command line.

This function is a modified version of the helper function _count_args
from the bash-completion project.