Fix usage and reword digest command
Since commit 0d9092c (casync: add seeking in archives), the command
digest can take a second parameter that is optional when the first
parameter is either an index or an archive (not blob index).

This rewording is inspired on the stat command section that is section
the most up-to-date.

This commit fixes the help usage, the man page for commands digest and
a typo in an error message.