casync: propagate --log-level option to helpers
Currently, casync-http handles the option --log-level but casync does
not propagate its verbosity to its protocol helpers.

This commit propagates the --log-level to the protocol helpers as it is
done for option --rate-limit-bps.

Now, the debug messages from the helper casync-http are printed to
stderr if the --log-level is specified to casync.