casync-http: Factorize curl handle work
The way we use curl handle is that we create it once, and then re-use it
again and again, as it's more efficient than re-allocating a new one for
each request.

By looking at the code closely, it turns out that the setup of the curl
handle needs to be done only once, then afterwards we only need to
change the URL in order to re-use the handle.

So this commit brings two helper functions to reflect that:
- make_curl_easy_handle() does the init work and set all the options for
  the handle.
- configure_curl_easy_handle() does the things that are needed in order
  to re-use the handle. In effect, it only sets the URL.

Additionally, this commit introduces curl_easy_cleanupp, in order to use
automatic pointers.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>