casync-http: Introduce macros to set curl options
These macros aim to make setting curl options easier.

CURL_SETOPT_EASY() sets the option, and on failure it outputs a generic
error message with the name of the option that failed, and returns -EIO.

The CURL_SETOPT_EASY_CANFAIL() variant does not return, it only outputs
an error message.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>