casync-http: add max-active-chunks option
This commit introduces a new option --max-active-chunks=<MAX> that
limits the number of simultaneous chunk transfers from the remote.

The MAX number is the sum of:

    1. the number of chunks added to the cURL multi interface, and
    2. chunks downloaded and waiting to be sent to the remote.

It limits the number of chunks stored in memory that are ready to be
sent to the casync process; this limit the memory usage in the situation
where the network is faster than the pipe communication between the
helper and casync.