Support streaming a block index' content into a char device
A reasonable exemplary usecase for this is updating a UBI volume (its interface
to userspace is a char device): let a parent process open the device file and
prepare the resulting file descriptor for updating the UBI volume, then fork
'casync extract <BLOB_INDEX> -' with STDOUT being redirected to that file
descriptor.

Without this change the only working alternative today is to fork 'casync mkdev
<BLOB_INDEX>' and then to actively pipe the resulting block device's content
into the UBI volume's prepared character device which is uneconomic as the
whole data has to be shuffled around once by casync and a second time by the
parent process.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>